controller, service 생성

+++ back-end/lotto/src/main/java/mirea/on/lotto/repo/LottoRepository.java
@@ -3,5 +3,5 @@ package mirea.on.lotto.repo;
 import mirea.on.lotto.entity.Lotto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LottoRepository extends JpaRepository<Lotto, Integer> {
+public interface LottoRepository extends JpaRepository<Lotto, Long> {
 }
